create common types in client
