+++ client/src/common/types.ts
@@ -1,0 +1,50 @@
+import { TConversation, TPreset } from 'librechat-data-provider';
+
+export type TSetOption = (param: number | string) => (newValue: number | string | boolean) => void;
+export type TSetExample = (
+  i: number,
+  type: string,
+  newValue: number | string | boolean | null,
+) => void;
+
+export enum ESide {
+  Top = 'top',
+  Right = 'right',
+  Bottom = 'bottom',
+  Left = 'left',
+}
+
+export type TBaseSettingsProps = {
+  conversation: TConversation | TPreset | null;
+  className?: string;
+  isPreset?: boolean;
+  readonly?: boolean;
+};
+
+export type TSettingsProps = TBaseSettingsProps & {
+  setOption: TSetOption;
+};
+
+export type TModels = {
+  models: string[];
+};
+
+export type TModelSelectProps = TSettingsProps & TModels;
+
+export type TEditPresetProps = {
+  open: boolean;
+  onOpenChange: React.Dispatch<React.SetStateAction<boolean>>;
+  preset: TPreset;
+  title?: string;
+};
+
+export type TSetOptionsPayload = {
+  setOption: TSetOption;
+  setExample: TSetExample;
+  addExample: () => void;
+  removeExample: () => void;
+  setAgentOption: TSetOption;
+  getConversation: () => TConversation | TPreset | null;
+  checkPluginSelection: (value: string) => boolean;
+  setTools: (newValue: string) => void;
+};
